+++ python/dalex/NEWS.md
@@ -1,8 +1,13 @@
 dalex (development)
 ----------------------------------------------------------------
 
+### bug fixes
+
+* `ModelPerformance.plot` now uses a drwhy color palette
+
 ### features
 
+* added the `ResidualDiagnostics` object with a `plot` method
 * added `model_diagnostics` method to the `Explainer`, which performs residual diagnostics
 * added `predict_surrogate` method to the `Explainer`, which is a wrapper for the `lime`
  tabular explanation from the [lime](https://github.com/marcotcr/lime) package